Rename test_util.rs to value.rs
