+++ examples/app-state/src/lib.rs
@@ -1,8 +1,7 @@
 mod app;
 mod drive;
 mod state;
-#[cfg(test)]
-mod test_utils;
+mod value;
 
 pub use app::app;
 #[cfg_browser(true)]
